此次提交更新了订单系统功能，完善了用户提交的逻辑。点击提交时模态框中js代码未能正确执行

+++ src/main/java/com/hubuteam/ordersystem/pojo/Merchant.java
@@ -25,7 +25,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Merchant {
-    private Long merchantId;
+    private int merchantId;
     private String merchantName;
     private String password;
     private String email;
